% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{CPIAggr}
\alias{CPIAggr}
\title{The CPIAggr <- function(fvcAppLanguage) ...
\itemize{
\item Receives a language argument
\item Creates global variables and updates loaded dataframes
\item Downloads all data from CODR tables 18100004 and 18100007
\item Defines reusable UI-related functions
\item Defines the ui function, which creates and positions UI objects and creates JS functions
\item Defines the server function, which receives user input and displays results
\item Calls shinyApp(ui, server)
}}
\usage{
CPIAggr(fvcAppLanguage)
}
\arguments{
\item{fvcAppLanguage}{The dusokay language, 'en' or 'fr'. All other values default to 'en'.}
}
\description{
The CPIAggr <- function(fvcAppLanguage) ...
\itemize{
\item Receives a language argument
\item Creates global variables and updates loaded dataframes
\item Downloads all data from CODR tables 18100004 and 18100007
\item Defines reusable UI-related functions
\item Defines the ui function, which creates and positions UI objects and creates JS functions
\item Defines the server function, which receives user input and displays results
\item Calls shinyApp(ui, server)
}
}
